rename: mock폴더 관련 구조 변경
