+++ src/mock/mockData/mockTemplateImage.ts
